refactor(maven): Extract util functions from datasource index
