+++ lib/datasource/maven/types.ts
@@ -1,0 +1,6 @@
+export interface MavenDependency {
+  display: string;
+  group?: string;
+  name?: string;
+  dependencyUrl: string;
+}
